refactor(policydb): Segregation of PolicyAssignmentController servicer
