+++ lte/cloud/go/services/policydb/policydb/main.go
@@ -20,7 +20,7 @@ import (
 	"magma/lte/cloud/go/protos"
 	"magma/lte/cloud/go/services/policydb"
 	"magma/lte/cloud/go/services/policydb/obsidian/handlers"
-	"magma/lte/cloud/go/services/policydb/servicers"
+	policydb_servicer "magma/lte/cloud/go/services/policydb/servicers/southbound"
 	"magma/orc8r/cloud/go/obsidian"
 	"magma/orc8r/cloud/go/obsidian/swagger"
 	swagger_protos "magma/orc8r/cloud/go/obsidian/swagger/protos"
@@ -33,7 +33,7 @@ func main() {
 	if err != nil {
 		glog.Fatalf("Error creating service: %s", err)
 	}
-	assignmentServicer := servicers.NewPolicyAssignmentServer()
+	assignmentServicer := policydb_servicer.NewPolicyAssignmentServer()
 	protos.RegisterPolicyAssignmentControllerServer(srv.GrpcServer, assignmentServicer)
 
 	swagger_protos.RegisterSwaggerSpecServer(srv.GrpcServer, swagger.NewSpecServicerFromFile(policydb.ServiceName))
